refactor: use fixtures
